requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=8.3.0 